modifications to file naming - prepping for nextflow

+++ src/scsilhouette/compute.py
@@ -1,5 +1,6 @@
 # src/scsilhouette/compute.py
 
+import os
 from pathlib import Path
 import pandas as pd
 import scanpy as sc
@@ -17,14 +18,17 @@ def run_silhouette(
     pca_components: int = None,
     save_scores: bool = False,
     save_cluster_summary: bool = False,
-    save_csv: bool = False,
-    show_obs: bool = False,
+    save_annotation: bool = False,
+    show_annotation: bool = False,
 ):
     from sklearn.metrics import silhouette_samples
     from sklearn.decomposition import PCA
     import numpy as np
 
     adata = sc.read(h5ad_path)
+    prefix = h5ad_path.name.replace(".h5ad", "")
+
+    print(f" will save output with {prefix}")
     print(f"[PASS] Loaded AnnData with {adata.shape}")
 
     # Handle gene subsetting if binary genes used
@@ -57,10 +61,14 @@ def run_silhouette(
     adata.obs["silhouette_score"] = scores
 
     # Save outputs
-    output_dir.mkdir(parents=True, exist_ok=True)
+    output_dir = Path(output_dir)
+
+    if not os.path.exists(output_dir):
+        raise FileNotFoundError(
+            f"[FAIL] Output directory '{output_dir}' does not exist. Please create it or use Nextflow publishDir.")
 
     if save_scores:
-        scores_path = output_dir / f"{label_key}_silhouette_scores.csv"
+        scores_path = output_dir / f"silhouette_scores_{prefix}_{label_key}_{embedding_key}_{metric}.csv"
         adata.obs[[label_key, "silhouette_score"]].to_csv(scores_path)
         print(f"[PASS] Saved silhouette scores to {scores_path}")
 
@@ -69,18 +77,20 @@ def run_silhouette(
             adata.obs.groupby(label_key)
             .agg(mean_silhouette=("silhouette_score", "mean"),
                  std_silhouette=("silhouette_score", "std"),
-                 median_silhouette=("silhouette_score", "median"))
+                 median_silhouette=("silhouette_score", "median"),
+                 count=("silhouette_score","count")))
+            .reset_index()
         )
-        summary_path = output_dir / f"{label_key}_cluster_summary.csv"
+        summary_path = output_dir / f"cluster_summary_{prefix}_{label_key}_{embedding_key}_{metric}.csv"
         summary_df.to_csv(summary_path)
         print(f"[PASS] Saved cluster summary to {summary_path}")
 
-    if save_csv:
-        obs_path = output_dir / f"{label_key}_obs.csv"
-        adata.obs.to_csv(obs_path)
-        print(f"[PASS] Saved obs to {obs_path}")
+    if save_annotation:
+        annotation_path = output_dir / f"annotation__{prefix}_{label_key}_{embedding_key}_{metric}.csv"
+        adata.obs.head().to_csv(annotation_path)
+        print(f"[PASS] Saved annotation to {obs_path}")
 
-    if show_obs:
+    if show_annotation:
         print(adata.obs.head())
 
     return adata
